rename observation module to obs (to avoid conflict with function of same name)
