+++ modelskill/model/dfsu.py
@@ -12,7 +12,7 @@ from ..quantity import Quantity
 from ..utils import _get_idx
 from .point import PointModelResult
 from .track import TrackModelResult
-from ..observation import Observation, PointObservation, TrackObservation
+from ..obs import Observation, PointObservation, TrackObservation
 
 
 class DfsuModelResult(SpatialField):
