+++ modelskill/model/track.py
@@ -3,7 +3,7 @@ from typing import Optional, Sequence
 
 import xarray as xr
 
-from ..observation import TrackObservation
+from ..obs import TrackObservation
 from ..types import TrackType
 from ..quantity import Quantity
 from ..timeseries import TimeSeries, _parse_track_input
