+++ tests/test_eum.py
@@ -1,8 +1,7 @@
-from modelskill.observation import unit_display_name
+from modelskill.obs import unit_display_name
 
 
 def test_units_display_name():
-
     assert unit_display_name("meter") == "m"
     assert unit_display_name("meter_per_sec") == "m/s"
     assert unit_display_name("second") == "s"
